finish step 7 and revise the project structure
