+++ go.mod
@@ -1,15 +1,15 @@
-module todo-app
+module practicum_final_project
 
 go 1.22.4
 
 require (
-	github.com/jmoiron/sqlx v1.4.0
-	github.com/mattn/go-sqlite3 v1.14.24
-	github.com/stretchr/testify v1.10.0
+    github.com/jmoiron/sqlx v1.4.0
+    github.com/mattn/go-sqlite3 v1.14.24
+    github.com/stretchr/testify v1.10.0
 )
 
 require (
-	github.com/davecgh/go-spew v1.1.1 // indirect
-	github.com/pmezard/go-difflib v1.0.0 // indirect
-	gopkg.in/yaml.v3 v3.0.1 // indirect
+    github.com/davecgh/go-spew v1.1.1 // indirect
+    github.com/pmezard/go-difflib v1.0.0 // indirect
+    gopkg.in/yaml.v3 v3.0.1 // indirect
 )
