+++ handlers/db.go
@@ -1,0 +1,11 @@
+package handlers
+
+import (
+	"database/sql"
+)
+
+var db *sql.DB
+
+func InitDB(database *sql.DB) {
+	db = database
+}
